child: fix missing backslash before & in tex heading
Need to use \\\& rather than just \&,
because tex_heading in Cite this page does escaping
for the special literals \~, \{, and \}.
Functional literal escaping (for backslash and pipe) is not performed,
neither is it for the other escapes (\&, \<, \>, etc.).
Functional literal escaping cannot be done
because Cite this page is an element with pipe-delimited arguments.
To allow for inserting a literal pipe, \| must be able to work,
and so backslashes cannot be escaped either.