Refactor <body> margin and padding
Here I set all body margins to zero
and allocate everything to body padding.
Previously the top and sides were styled using padding,
and the bottom was styled using margin.
It makes more sense for everything to be controlled by padding,
and nothing to be controlled by margin (i.e. all or nothing).