Convert italics from CCH to CMD
Here only translator-supplied italics are intended to be converted.
Book-title italics will be inadvertently converted too;
change those to _{content}_ manually.

Find:
   \{([a-z ]+)\}
Replace with:
   .[\1]

There is a single leading space in both strings,
and the content must consist of letters and spaces only.
This was the simplest way to prevent unwanted replacements
of curly bracketed strings inside mathematical equations.