Convert colour opening tags from CCH to CMD
Find:
  <:([a-z])>
Replace with:
  [\1/

Note that the CMD syntax does not allow nesting,
and that some choices of letter in CCH have no meaning in CMD.

Also note that the CMD syntax allows whitespace
in content delimited by `[c/ /]`,
and this should be used for greater readability.
The replacement here does not insert any whitespace.