Use explicit class "start"/"end" for markers
Whereas previously the markers for start/end
in Cynic's 7.30 were of class "marker",
with first-child and last-child selectors
to determine if it marked the start/end of an excerpt/interview,
now explicit classes "start" and "end" are used to this end.
This is because markers will be used
to mark the end of Preface/Volumes in Sun Tzu,
and in this case the marker will not be the last child.
Therefore an explicit "end" class is required.
And given that an "end" class will be used,
I should also make a "start" class for symmetry.