Initial commit for branch 'inclusions-order'
As of conway-markdown/conway-markdown@618dc637,
regex and ordinary replacements now have an optional flag
for specifying exactly when they should be applied.
Hopefully this can be used to eliminate the strict load order
for resources/inclusions/syntax/*.cmd.
More generally, the strategy is to use flags
so that replacements are applied as late as possible
('just before' the earliest syntax used in the replacement).
God save!