Use w flag for auto Chinese lang
Specifically two after processing headings,
so that heading ids containing Chinese characters
need not be wrapped in CMD literals
in order to prevent unwanted lang spans.
The heading permalinks regex in general.cmd
will be modified to be one after processing headings,
so that the id is placeholder protected
before it is used in the href element of the permalink.