Lexicon: add all radicals (in comments)
May as well do all of this in one go. Uncomment as appropriate.
Note that the characters to be used for the radicals
are in the Unicode block 'Kangxi Radicals' (U+2F00 through U+2FDF).
While they look the same as the characters in the main CJK blocks,
they have different code points.