Fix rod numeral SVG height in Sun Tzu I P7 table
(For some reason the rod numeral SVGs in the table
are rendered at height ~150 px.)
This effectively reverts 756a1c08 for /sun-tzu/i/7.
That commit was a temporary fix for Sun Tzu square root SVG height
on DuckDuckGo, probably inheriting a Chromium bug,
and ought to be fully reverted once the bug is fixed.