Initial commit for branch 'archive'
This repository is quite large (48M according to du) now,
mainly due to the PDFs in `/manuscripts/`
and especially due to those which have been re-scanned,
whose old versions permanently bloat the repository
despite no longer existing at the tip.
Therefore I have decided to remove all of the PDFs
and also some large images, and make a new repository
not containing these historic large files,
in order to save bandwidth without having to shallow clone
all the time (and fiddle around with `set-branches`).
The old repository will be moved to
<https://github.com/yawnoc/yawnoc.github.io-archive-1>
and the existing links to the PDFs pointed there.